==> real/grun_2016/03_qumi.Rmd <==
---
title: "Quasi-UMIs for Grun 2016"
author: "Will Townes"
output: html_document
---

Estimate poisson-lomax, poisson-lognormal, and negative binomial (poisson-gamma) parameters for each cell

```{r}
suppressPackageStartupMessages(library(SingleCellExperiment))
library(ggplot2); theme_set(theme_bw())
source("./algs/quminorm.R") #also loads nblomax.R
source("./algs/poilog.R")
fp<-file.path
bp<-"./real/grun_2016"
plt_dir<-fp(bp,"results/fig")
if(!dir.exists(plt_dir)) dir.create(plt_dir,recursive=TRUE)
```

data loading

```{r}
sce<-readRDS(fp(bp,"data/01_sce_all_genes_all_cells.rds"))
m<-counts(sce)
```

## Exploratory log-log plots

PMF of random cell, compare poisson-lomax with poisson-lognormal

```{r}
set.seed(333)
i<-sample.int(ncol(m),size=1)
colnames(m)[i] #GSM2142268_ACTGATCG
x<-m[,i]
(th<-poilog_mle(x))
(th2<-plomax_mle(x,quadpts=1000))
(th3<-nb_mle(x))
res<-lpmf(x,doplot=FALSE)
xmax<-expm1(5.3)

fpoilog<-function(t){
  sads::dpoilog(floor(exp(t)),mu=th[1],sig=th[2],log=TRUE)
}
fplomax<-function(t){
  dplomax(floor(exp(t)),tail=th2[1],scale=th2[2],log=TRUE,quadpts=10000)
}
fnb<-function(t){
  dnbinom(floor(exp(t)),size=th3[1],mu=th3[2],log=TRUE)
}

pdf(file=fp(plt_dir,"lpmf_llcurve_grun.pdf"),width=4,height=4)
marpar<-par("mar"); par(mar=c(4,4,.5,.5))
lpmf_xtra(res,xlab="UMI count",ylab="log(density)",connect=FALSE,ylim=c(-16,0),xlim=c(-log(2),log(xmax)),cex=2,lwd=3)
curve(fpoilog,from=log(1),to=log(xmax),add=TRUE,col="blue",lwd=3)
curve(fplomax,from=log(1),to=log(xmax),add=TRUE,col="green",lwd=3,lty=2)
curve(fnb,from=log(1),to=log(xmax),add=TRUE,col="red",lwd=3,lty=3)
points(rep(-log(2),3),c(fpoilog(-Inf),fplomax(-Inf),fnb(-Inf)),col=c("blue","green","red"),pch=c(1,1,19),cex=c(1.6,.9,.4),lwd=3)
legend("topright",c("Poisson-lognormal","Poisson-Lomax","negative binomial"),col=c("blue","green","red"),lty=c(1,2,3),lwd=3)
par(mar=marpar)
dev.off()
```

## Compute MLEs for all cells 

`time Rscript ./real/util/mles_qumi.R ./real/grun_2016 data/01_sce_all_genes_all_cells.rds` 
or 
`sbatch -J grun ./real/util/mles_qumi.slurm ./real/grun_2016 data/01_sce_all_genes_all_cells.rds`

## Goodness of fit 

### BIC

```{r}
pd<-read.table(fp(bp,"results/mle_merged.txt"),header=TRUE)
ln<-subset(pd,method=="poilog")
lx<-subset(pd,method=="poilomax")
nb<-subset(pd,method=="negbinom")
#when is poilog best?
mean(ln$bic<pmin(lx$bic,nb$bic))
#when is poisson-lomax best?
mean(lx$bic<pmin(ln$bic,nb$bic))
#nb best?
mean(nb$bic<pmin(ln$bic,lx$bic))
with(pd,boxplot(bic~method,log="y")) #very small differences
```

Based on BIC, in 39% of cells, Poisson-lognormal fits best. In 17% of cells Poisson-lomax fits best. In 44% of cells, negative binomial fits best. The overall differences between methods are very small according to BIC.

### Predictive check based on simulated maxima

See *real/zhao_2019/02_qumi.Rmd* for details.

```{r}
cpmax<-read.table(fp(bp,"results/compare_maxima.txt"),header=TRUE)
cpmax$log_err<-log(cpmax$sim_max)-log(cpmax$true_max)
ggplot(cpmax,aes(x=log_err))+geom_histogram()+facet_wrap(~method,scales="free")
```

Negative binomial underestimates the maximum, Poisson-lognormal is well-calibrated, and Poisson-Lomax overestimates.

## Distributions of parameter estimates (MLEs)

```{r}
pd<-read.table(fp(bp,"results/mle_merged.txt"),header=TRUE)
sce<-readRDS(fp(bp,"data/01_sce_all_genes_all_cells.rds"))
sce<-sce[,as.character(unique(pd$cell))]
cm<-colData(sce)
m<-counts(sce)
cm$cell<-rownames(cm)
pd<-merge(pd,as.data.frame(cm),by="cell")
```

```{r}
ggplot(pd,aes(x=shape))+geom_histogram(bins=100)+facet_wrap(~method,scales="free")+scale_x_log10()
pd$celltype<-sub("cell type: ","",pd$characteristics_ch1)
ggplot(pd,aes(x=celltype,y=shape))+geom_boxplot()+facet_grid(~method,scales="free_x")+coord_flip()+scale_y_log10()
```

For neg binom, the size parameter ranges from .005 to .1 with .02 a typical value. For poisson-lognormal the sig parameter ranges from 2.0-3.2 with median 2.7 and mean 2.6. For poisson-lomax the tail parameter distribution is somewhat bimodal with a major peak at .8 and a minor peak at 2.0.

There does appear to be some cell-type specificity to the shape parameter estimates.

### Comparing fitted MLEs to summary statistics

```{r}
# z<-log10(colSums(m))
# plot(z,ln$sig)
# plot(z,ln$mu)
# plot(z,lx$tail)
# plot(z,log(lx$scale))
```
